migrate

+++ database/migrations/2024_02_02_152212_remove_user_id_from_tasks.php
@@ -13,8 +13,10 @@ return new class extends Migration {
     public function up()
     {
        Schema::table('tasks', function (Blueprint $table) {
+        if(Schema::hasColumn('tasks', 'user_id')){
             $table->dropForeign(['user_id']);
             $table->dropColumn('user_id');
+        }
         });
     }
